requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
Pillow
python-dotenv
djangorestframework-jwt
psycopg2-binary
slugify